Move shared_ptr<Event> into execute task
