+++ src/parallel/pipeline_initialize_event.cpp
@@ -11,11 +11,10 @@ PipelineInitializeEvent::PipelineInitializeEvent(shared_ptr<Pipeline> pipeline_p
 class PipelineInitializeTask : public ExecutorTask {
 public:
 	explicit PipelineInitializeTask(Pipeline &pipeline_p, shared_ptr<Event> event_p)
-	    : ExecutorTask(pipeline_p.executor), pipeline(pipeline_p), event(std::move(event_p)) {
+	    : ExecutorTask(pipeline_p.executor, std::move(event_p)), pipeline(pipeline_p) {
 	}
 
 	Pipeline &pipeline;
-	shared_ptr<Event> event;
 
 public:
 	TaskExecutionResult ExecuteTask(TaskExecutionMode mode) override {
